+++ src/AtomUI.Desktop.Controls/Switch/ToggleSwitch.SemanticParts.cs
@@ -1,0 +1,20 @@
+using AtomUI.Theme;
+using Avalonia.Controls.Presenters;
+using Avalonia.Controls.Primitives;
+
+namespace AtomUI.Desktop.Controls;
+
+[SemanticPart(
+    "content",
+    SelectorClass = "semantic-content",
+    ContractType = typeof(ContentPresenter),
+    Cardinality = SemanticPartCardinality.Multiple,
+    Since = "6.0")]
+[SemanticPart(
+    "indicator",
+    SelectorClass = "semantic-indicator",
+    ContractType = typeof(TemplatedControl),
+    Since = "6.0")]
+public partial class ToggleSwitch
+{
+}