+++ src/AtomUI.Desktop.Controls/Switch/ToggleSwitch.cs
@@ -2,7 +2,7 @@ using AtomUI.Controls;
 
 namespace AtomUI.Desktop.Controls;
 
-public class ToggleSwitch : AbstractToggleSwitch
+public partial class ToggleSwitch : AbstractToggleSwitch
 {
     public ToggleSwitch()
     {